requirements.txt
adal==1.2.6
azure-common==1.1.25
azure-mgmt-monitor==0.11.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.1.0
azure-mgmt-subscription==0.6.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
isodate==0.6.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.4
